Improve styling by moving bunq promo details to account page
