+++ src/react/Locales/en.json
@@ -425,7 +425,7 @@
 	"Something went wrong while trying to authorize the client": "Something went wrong while trying to authorize the client",
 	"Something went wrong!": "Something went wrong!",
 	"Spent amount monthly": "Spent amount monthly",
-	"Spent of monthly budget": "Spent of monthly budget",
+	"Spent this month": "Spent this month",
 	"Split request": "Split request",
 	"Start amount": "Start amount",
 	"Start amount can't be more than the goal amount": "Start amount can't be more than the goal amount",
